More code to Add a to do list and save it as a file
