+++ app/src/main/java/ToDo/App.java
@@ -2,12 +2,14 @@
  * This Java source file was generated by the Gradle 'init' task.
  */
 package ToDo;
-import java.io.File;
+
 import java.util.Scanner;
+
 public class App {
     public String getGreeting() {
         return "ToDo List Manager";
     }
+
     public static void main(String[] args) {
         String input = mainScreenDisplay();
         switch (input) {
@@ -17,16 +19,21 @@ public class App {
                 break;
             case "2":
 //                Display ToDo List
-                System.out.println(input + " was returned");
+                System.out.println("Display contents of ToDoList.txt file");
                 try {
-                    displayToDoList();
+                    Display.displayToDoList();
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
                 break;
             case "3":
 //                Add a ToDo List
-                System.out.println(input + " was returned");
+                //System.out.println(input + " was returned");
+                try {
+                    Add.addToDoList();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
                 break;
             case "4":
 //                Edit ToDo List
@@ -46,6 +53,7 @@ public class App {
                 break;
         }
     }
+
     public static String mainScreenDisplay() {
         Scanner s = new Scanner(System.in);
         System.out.println("                          Main Menu");
@@ -54,7 +62,7 @@ public class App {
         System.out.println("2.  Display ToDo List");
         System.out.println("3.  Add a ToDo List");
         System.out.println("4.  Edit ToDo List");
-        System.out.println("5.  Delete ToDo List");2
+        System.out.println("5.  Delete ToDo List");
         System.out.println("6.  Save ToDo List");
         System.out.println("7.  Exit");
         System.out.println("\n");
@@ -63,13 +71,4 @@ public class App {
         return u;
     }
 
-    public static void displayToDoList() throws Exception {
-        File file = new File("C:\\Users\\k330620\\Documents\\Java_class\\JavaGettingStarted_ToDoList\\app\\src\\main\\java\\ToDo\\ToDoList.txt");
-        Scanner sc = new Scanner(file);
-        StringBuffer sb = new StringBuffer();
-        while (sc.hasNext()) {
-            sb.append(" " + sc.nextLine());
-        }
-        System.out.println(sb);
-    }
 }
