+++ app/src/main/java/ToDo/ToDoList.txt
@@ -1,5 +1,6 @@
-ToDo List:
+MELISSA'S TODO LIST:
 
-1. Task A,
-2. Task B,
-3. Task C.
+1. Remove old siding.
+2. Put up sheeting.
+3. Attach Tyvek wrap.
+4. Install new siding.
